requirements.txt
Django==1.8.6
djangorestframework==3.3.1
simplejson==3.8.1
MySQL-python==1.2.5
grequests==0.3.0
requests==2.9.1
setproctitle==1.1.9
prospector==0.11
pyroma==1.8.3
mock==1.3.0
python-dateutil==2.6.0
pytz==2016.10

